requirements.txt
alembic==1.16.5
bcrypt==4.3.0
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
colorama==0.4.6
dnspython==2.8.0
email-validator
eventlet==0.33.3
Flask==2.3.3
Flask-Bcrypt==1.0.1
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Mail==0.9.1
Flask-Migrate==4.0.5
flask-marshmallow==1.3.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.0.5
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-auth==2.40.3
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.62.3
gunicorn==21.2.0
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==3.19.0
packaging==25.0
proto-plus==1.26.1
protobuf==4.25.8
psycopg-binary
pydantic
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
python-dotenv==1.0.0
python-engineio==4.12.2
python-socketio==5.8.0
requests==2.31.0
rsa==4.9.1
simple-websocket==1.1.0
six==1.17.0
SQLAlchemy==2.0.43
tqdm==4.67.1
typing_extensions==4.15.0
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
pytest==8.4.2
pytest-flask==1.3.0
marshmallow-sqlalchemy==0.29.0
